メニューの実装(スマホ) (#147)
* fill-rule -> fillRule

* Spメニューを開くためのボタンを実装

* SpMenuで表示する要素を実装

* SpMenu結合

* SpMenu配置

* lg mdではhidden

* CTAAreaのボタンにpeatixのリンクを追加

---------
